eclone: Turned clone_flags into a 64bits variable in do_fork()
sys_eclone() allows the user to pass a 64bits clone flags, do_fork()
handles it properly.

+++ arch/alpha/kernel/process.c
@@ -271,7 +271,7 @@ alpha_vfork(struct pt_regs *regs)
  */
 
 int
-copy_thread(unsigned long clone_flags, unsigned long usp,
+copy_thread(unsigned long long clone_flags, unsigned long usp,
 	    unsigned long unused,
 	    struct task_struct * p, struct pt_regs * regs)
 {
