+++ arch/blackfin/kernel/process.c
@@ -189,7 +189,7 @@ asmlinkage int bfin_clone(struct pt_regs *regs)
 }
 
 int
-copy_thread(unsigned long clone_flags,
+copy_thread(unsigned long long clone_flags,
 	    unsigned long usp, unsigned long topstk,
 	    struct task_struct *p, struct pt_regs *regs)
 {
