+++ arch/score/kernel/process.c
@@ -85,7 +85,7 @@ void flush_thread(void) {}
 /*
  * set up the kernel stack and exception frames for a new process
  */
-int copy_thread(unsigned long clone_flags, unsigned long usp,
+int copy_thread(unsigned long long clone_flags, unsigned long usp,
 		unsigned long unused,
 		struct task_struct *p, struct pt_regs *regs)
 {
