+++ include/linux/sem.h
@@ -144,11 +144,11 @@ struct sysv_sem {
 
 #ifdef CONFIG_SYSVIPC
 
-extern int copy_semundo(unsigned long clone_flags, struct task_struct *tsk);
+extern int copy_semundo(unsigned long long clone_flags, struct task_struct *tsk);
 extern void exit_sem(struct task_struct *tsk);
 
 #else
-static inline int copy_semundo(unsigned long clone_flags, struct task_struct *tsk)
+static inline int copy_semundo(unsigned long long clone_flags, struct task_struct *tsk)
 {
 	return 0;
 }
