+++ kernel/cred.c
@@ -380,7 +380,7 @@ struct cred *prepare_exec_creds(void)
  * The new process gets the current process's subjective credentials as its
  * objective and subjective credentials
  */
-int copy_creds(struct task_struct *p, unsigned long clone_flags)
+int copy_creds(struct task_struct *p, unsigned long long clone_flags)
 {
 #ifdef CONFIG_KEYS
 	struct thread_group_cred *tgcred;
