+++ kernel/sched.c
@@ -2449,7 +2449,7 @@ static void __sched_fork(struct task_struct *p)
 /*
  * fork()/clone()-time setup:
  */
-void sched_fork(struct task_struct *p, int clone_flags)
+void sched_fork(struct task_struct *p, unsigned long long clone_flags)
 {
 	int cpu = get_cpu();
 
@@ -2528,7 +2528,7 @@ void sched_fork(struct task_struct *p, int clone_flags)
  * that must be done for every newly created context, then puts the task
  * on the runqueue and wakes it.
  */
-void wake_up_new_task(struct task_struct *p, unsigned long clone_flags)
+void wake_up_new_task(struct task_struct *p, unsigned long long clone_flags)
 {
 	unsigned long flags;
 	struct rq *rq;
